Important message about Fedora 32.
This README was only updated to indicate how graphic problems may appear if trying to install VRS on Fedora 32.  To be clear, the problem is not necessarily with Fedora 32, but with the fact that Fedora 32 installs a version of Mono that is not less than 6.4.  Running Mono version 6.4 or higher will cause graphic problems with VRS.