New feature:  VRS watchdog script.
This version (11.1) just has a few cosmetic updates from the previous version (11.1).

It is important to know that 11.0 introduced a new VRS watchdog script that now gets produced by this installation script.  The VRS watchdog script will allow the user to continually monitor VRS and to make sure it is running 24/7.