Information on VRS watchdog script
This version just has a few cosmetic updates from the previous version.

But, it is important to know that 11.0 introduced a new VRS watchdog script that now gets produced by this installation script.  The VRS watchdog script will allow the user to continually monitor VRS and to make sure it is running 24/7.

There is also an update that this script does install correctly on the latest versions of a few Linux distributions:  Debian 10.10, Linux Mint 20.2 and openSUSE 15.3.